improve execute function and use scope function
