+++ src/main/kotlin/Main.kt
@@ -1,7 +1,6 @@
 import dataSource.CsvDataSource
 import dataSource.utils.CsvParser
 import interactor.*
-import java.lang.System.exit
 import kotlin.system.exitProcess
 
 val csvParser = CsvParser()
@@ -159,7 +158,7 @@ fun getChooseOptionNumber(): Int {
 Side of Get Data element -> Instance Zone
  *********************************************************/
 fun getTop10CountriesWithHighTaxOnCarbonatedDrinks(limit: Int) {
-    val getTop10CountriesWithHighTaxOnCarbonatedDrinks = GetTop10CountriesWithHighTaxOnCarbonatedDrinks(dataSource)
+    val getTop10CountriesWithHighTaxOnCarbonatedDrinks = GetTopCountriesWithHighTaxOnCarbonatedDrinks(dataSource)
     println(getTop10CountriesWithHighTaxOnCarbonatedDrinks.execute(10, data))
 }
