--- src/main/kotlin/interactor/GetTop10CountriesWithHighTaxOnCarbonatedDrinks.kt
@@ -1,23 +1,0 @@
-package interactor
-import model.DrinksPrices
-import model.CityEntity
-import java.util.*
-
-class GetTop10CountriesWithHighTaxOnCarbonatedDrinks ( private val dataSource: CostOfLivingDataSource){
-    fun execute(limit:Int,cities:List<CityEntity>): List<Pair<String, Float?>> {
-        return cities
-            .asSequence()
-            .filter(::excludeNullDrinkPriceAndNegativePriceAndLowQualityData)
-            .distinctBy { it.country }
-            .sortedByDescending { it.drinksPrices.cokePepsiAThirdOfLiterBottleInRestaurants }
-            .take(limit)
-            .map { Pair(it.country ,it.drinksPrices.cokePepsiAThirdOfLiterBottleInRestaurants)}
-            .toList()
-    }
-    private fun excludeNullDrinkPriceAndNegativePriceAndLowQualityData(city: CityEntity): Boolean {
-        return  city.drinksPrices.cokePepsiAThirdOfLiterBottleInRestaurants != null
-                && city.drinksPrices.cokePepsiAThirdOfLiterBottleInRestaurants  > 0
-                && city.country!= null
-                && city.dataQuality
-    }
-}
